Rename acclimatise module and package to aclimatise
