+++ aclimatise/usage_parser/parser.py
@@ -3,10 +3,10 @@ from typing import List
 
 from pyparsing import *
 
-from acclimatise.flag_parser.elements import customIndentedBlock, description_line
-from acclimatise.parser import IndentCheckpoint, IndentParserMixin
-from acclimatise.usage_parser.elements import usage_example
-from acclimatise.usage_parser.model import UsageElement, UsageInstance
+from aclimatise.flag_parser.elements import customIndentedBlock, description_line
+from aclimatise.parser import IndentCheckpoint, IndentParserMixin
+from aclimatise.usage_parser.elements import usage_example
+from aclimatise.usage_parser.model import UsageElement, UsageInstance
 
 from .elements import *
