+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = acclimatise
+name = aclimatise
 version = 1.2.0
 description = Acclimatise is a Python library and command-line utility for parsing the help output of a command-line tool and then outputting a description of the tool in a more structured format
 long_description = file: README.rst
